requirements.txt
flask
napalm
requests
pyyaml